Added Movies

+++ db/migrate/20211111115333_create_movies.rb
@@ -4,7 +4,7 @@ class CreateMovies < ActiveRecord::Migration[6.0]
       t.string :title
       t.text :overview
       t.string :poster_url
-      t.decimal :rating
+      t.float :rating
 
       t.timestamps
     end
